+++ service/interceptor/paramcheck/client.go
@@ -25,9 +25,16 @@ import (
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
 	api "github.com/polarismesh/polaris/common/api/v1"
+	"github.com/polarismesh/polaris/common/model"
 	"github.com/polarismesh/polaris/common/utils"
 	"github.com/polarismesh/polaris/service"
 )
+
+// GetPrometheusTargets implements service.DiscoverServer.
+func (svr *Server) GetPrometheusTargets(ctx context.Context,
+	query map[string]string) *model.PrometheusDiscoveryResponse {
+	return svr.nextSvr.GetPrometheusTargets(ctx, query)
+}
 
 // RegisterInstance create one instance by client
 func (s *Server) RegisterInstance(ctx context.Context, req *apiservice.Instance) *apiservice.Response {
